+++ Docs/Changes.md
@@ -1,4 +1,5 @@
 # 0.9.14
+  * Change: Multiple runtime error messages are corrected.
   * Fixed: _Incorrect exception message for custom types for `IndexOutOfRange`_ ([Issue #253](https://github.com/vorov2/dyalect/issues/253)).
   
 # 0.9.13